refactor: make first draft of final structure
